show 8 templates on homepage instead of 10
Former-commit-id: bef5b876360f68ae03b661b8f32e22f28384b46e

+++ intermine/web/main/src/org/intermine/web/struts/BeginAction.java
@@ -43,7 +43,7 @@ import org.intermine.web.logic.session.SessionMethods;
 public class BeginAction extends InterMineAction
 {
     private static final Logger LOG = Logger.getLogger(TemplateTracker.class);
-    private static final Integer MAX_TEMPLATES = new Integer(10);
+    private static final Integer MAX_TEMPLATES = new Integer(8);
 
      /**
      * Either display the query builder or redirect to project.sitePrefix.
